Nicer behaviour from ack-skel-vc-grep when used in setup hook
Close #9, close #10.

Previously, if the user had ack-skel-vc-grep in the
ack-minibuffer-setup-hook and chose a directory without a vc project,
C-u C-u M-x ack would fail because ack-skel-vc-grep raised an error,
since it can't work under those circunstances. This commit makes it
give up in these situations and leave the (presumably ack) prompt
unchanged.

Additionally, if the user chose a subdirectory *within* a vc project,
C-u C-u M-x ack would surprisingly change the directory to the vc
project root.  While this is probably intended in interactive calls to
ack-skel-vc-grep, it is not when that function is in the
ack-minibuffer-setup-hook.

* ack.el (ack-skel-vc-grep): Overhaul function. Distinguish
between interactive and non-interactive uses.