Weird Minibuffer Display When Candidate Search Text Has Properties
When the candidate symbol or text to search for is copied from the
current buffer to the command line in the minibuffer, text properties
are copied, too. When the text has a big font size, is bold, etc. this
can affect the minibuffer display in aesthetically unpleasant ways.
This commit removes any text properties from the candidate symbol, so
that the command line in the minibuffer retains a uniform appearance.

* ack.el (ack-yank-symbol-at-point): remove text properties from
candidate symbol